:art: refactor : 챗봇 구조 수정 완료

+++ backend/src/main/java/org/example/b104/domain/chatbot/service/command/ChatBotTextCommand.java
@@ -7,10 +7,10 @@ import lombok.Data;
 @Builder
 public class ChatBotTextCommand {
 
-    private Long command_id;
+    private Integer command_id;
     private String message;
     private String bank;
     private String account;
-    private Long money;
-    private Long is_chatbot;
+    private Integer money;
+    private Integer is_chatbot;
 }
